% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgOut.R
\docType{package}
\name{pkgOut}
\alias{pkgOut}
\title{pkgOut for demonstration purpose of linking native routines}
\description{
A out-going package that exports a native routine
}
\author{
Jitao David Zhang <jitao_david.zhang@roche.com>
}
