% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgOut.R
\name{pkgOutFunc}
\alias{pkgOutFunc}
\title{A dummy function exported for demonstration purposes}
\usage{
pkgOutFunc(username)
}
\arguments{
\item{username}{Username in characters}
}
\value{
A number generated in C
}
\description{
A dummy function exported for demonstration purposes
}
